First Iteration, NegaMax with alpha-beta pruning
NegaMax is used along with alpha-beta pruning to search for the best performing sequence of moves, based on the evaluation function.

The evaluation function is a simple one that compares the total value of a player's pieces to their opponent's. It does not take into account piece positioning.

A transposition table is used to store previously computed board states and speed up the search algorithm.

Useful Links:
https://www.chessprogramming.org/Negamax
https://www.chessprogramming.org/Transposition_Table
https://www.chessprogramming.org/Alpha-Beta