Quiesce search, Iterative Deepening, Heuristic improvements
Quiesce search is added after the negamax search. The Quiesce search doesn't use the transposition table due to a lack of available tokens. I will need to shorten MyBot before this will be possible.

The heuristic function is improved to incorporate the number of legal moves the current player is able to make. Improving the heuristic function further should help speed up the the search algorithm, which will enable a deeper search of the decision tree.

Iterative Deepening is added, however, due to a limited search depth, it's improvements are minimal so far.