Replacing depth limit with Time Management
I replaced the hard-coded iterative deepening limit with a function based on the Timer.

The idea is that we can search further down the decision tree when there is less fanout, or less possible moves. However, determining the best amount of time to allocate per move is difficult, I will need to spend more time on it.

https://www.chessprogramming.org/Time_Management